Add files via upload
Moved parse_config() to separate module and fixed many cosmetics things. Also added empty docstrings to be completed.
